+++ src/compute/package_entry.ts
@@ -1,0 +1,4 @@
+import { ensureComputeArgv } from "./entry";
+
+process.argv = ensureComputeArgv(process.argv);
+await import("../server");